requirements.txt
google-cloud-pubsub>=2.0.0
pydantic>=2.4.0
six>=1.16.0
pytz>=2021.1
protobuf>=3.19.0
